CLEAN: Remove & Sort Code
