--- app/Http/Controllers/ProductController.php
@@ -1,243 +1,0 @@
-<?php
-
-namespace App\Http\Controllers;
-
-use Exception;
-use App\Models\Unit;
-use App\Models\Product;
-use App\Models\Category;
-use Illuminate\Http\Request;
-use PhpOffice\PhpSpreadsheet\IOFactory;
-use PhpOffice\PhpSpreadsheet\Writer\Xls;
-use Picqer\Barcode\BarcodeGeneratorHTML;
-use PhpOffice\PhpSpreadsheet\Spreadsheet;
-use App\Http\Requests\Product\StoreProductRequest;
-use App\Http\Requests\Product\UpdateProductRequest;
-
-class ProductController extends Controller
-{
-    public function index()
-    {
-        $products = Product::select('id', 'name')
-            ->limit(1)
-            ->get();
-
-        return view('products.index', [
-            'products' => $products,
-        ]);
-    }
-
-    public function create(Request $request)
-    {
-        $categories = Category::all();
-        $units = Unit::all();
-
-        if ($request->has('category'))
-        {
-            $categories = Category::whereSlug($request->get('category'))->get();
-        }
-
-        if ($request->has('unit'))
-        {
-            $units = Unit::whereSlug($request->get('unit'))->get();
-        }
-
-        return view('products.create', [
-            'categories' => $categories,
-            'units' => $units,
-        ]);
-    }
-
-    public function store(StoreProductRequest $request)
-    {
-//        dd($request);
-        $product = Product::create($request->all());
-
-        /**
-         * Handle upload image
-         */
-        if($request->hasFile('product_image')){
-            $file = $request->file('product_image');
-            $filename = hexdec(uniqid()).'.'.$file->getClientOriginalExtension();
-
-            $file->storeAs('products/', $filename, 'public');
-            $product->update([
-                'product_image' => $filename
-            ]);
-        }
-
-        return redirect()
-            ->route('products.index')
-            ->with('success', 'Product has been created!');
-    }
-
-    public function show(Product $product)
-    {
-        // Generate a barcode
-        $generator = new BarcodeGeneratorHTML();
-
-        $barcode = $generator->getBarcode($product->code, $generator::TYPE_CODE_128);
-
-        return view('products.show', [
-            'product' => $product,
-            'barcode' => $barcode,
-        ]);
-    }
-
-    public function edit(Product $product)
-    {
-        return view('products.edit', [
-            'categories' => Category::all(),
-            'units' => Unit::all(),
-            'product' => $product
-        ]);
-    }
-
-    public function update(UpdateProductRequest $request, Product $product)
-    {
-        $product->update($request->except('product_image'));
-
-        if($request->hasFile('product_image')){
-
-            // Delete Old Photo
-            if($product->product_image){
-                unlink(public_path('storage/products/') . $product->product_image);
-            }
-
-            // Prepare New Photo
-            $file = $request->file('product_image');
-            $fileName = hexdec(uniqid()).'.'.$file->getClientOriginalExtension();
-
-            // Store an image to Storage
-            $file->storeAs('products/', $fileName, 'public');
-
-            // Save DB
-            $product->update([
-                'product_image' => $fileName
-            ]);
-        }
-
-        return redirect()
-            ->route('products.index')
-            ->with('success', 'Product has been updated!');
-    }
-
-    public function destroy(Product $product)
-    {
-        /**
-         * Delete photo if exists.
-         */
-        if($product->product_image){
-            unlink(public_path('storage/products/') . $product->product_image);
-        }
-
-        $product->delete();
-
-        return redirect()
-            ->route('products.index')
-            ->with('success', 'Product has been deleted!');
-    }
-
-    public function import()
-    {
-        return view('products.import');
-    }
-
-    public function handleImport(Request $request)
-    {
-        $request->validate([
-            'file' => 'required|file|mimes:xls,xlsx',
-        ]);
-
-        $the_file = $request->file('file');
-
-        try{
-            $spreadsheet = IOFactory::load($the_file->getRealPath());
-            $sheet        = $spreadsheet->getActiveSheet();
-            $row_limit    = $sheet->getHighestDataRow();
-            $column_limit = $sheet->getHighestDataColumn();
-            $row_range    = range( 2, $row_limit );
-            $column_range = range( 'J', $column_limit );
-            $startcount = 2;
-            $data = array();
-            foreach ( $row_range as $row ) {
-                $data[] = [
-                    'product_name' => $sheet->getCell( 'A' . $row )->getValue(),
-                    'category_id' => $sheet->getCell( 'B' . $row )->getValue(),
-                    'unit_id' => $sheet->getCell( 'C' . $row )->getValue(),
-                    'product_code' => $sheet->getCell( 'D' . $row )->getValue(),
-                    'stock' => $sheet->getCell( 'E' . $row )->getValue(),
-                    'buying_price' => $sheet->getCell( 'F' . $row )->getValue(),
-                    'selling_price' =>$sheet->getCell( 'G' . $row )->getValue(),
-                    'product_image' =>$sheet->getCell( 'H' . $row )->getValue(),
-                ];
-                $startcount++;
-            }
-
-            Product::insert($data);
-
-        } catch (Exception $e) {
-            // $error_code = $e->errorInfo[1];
-            return redirect()
-                ->route('products.index')
-                ->with('error', 'There was a problem uploading the data!');
-        }
-
-        return redirect()
-            ->route('products.index')
-            ->with('success', 'Data product has been imported!');
-    }
-
-    function export()
-    {
-        $products = Product::all()->sortBy('product_name');
-
-        $product_array [] = array(
-            'Product Name',
-            'Category Id',
-            'Unit Id',
-            'Product Code',
-            'Stock',
-            'Buying Price',
-            'Selling Price',
-            'Product Image',
-        );
-
-        foreach($products as $product)
-        {
-            $product_array[] = array(
-                'Product Name' => $product->product_name,
-                'Category Id' => $product->category_id,
-                'Unit Id' => $product->unit_id,
-                'Product Code' => $product->product_code,
-                'Stock' => $product->stock,
-                'Buying Price' =>$product->buying_price,
-                'Selling Price' =>$product->selling_price,
-                'Product Image' => $product->product_image,
-            );
-        }
-
-        $this->exportExcel($product_array);
-    }
-
-    public function exportExcel($products)
-    {
-        ini_set('max_execution_time', 0);
-        ini_set('memory_limit', '4000M');
-
-        try {
-            $spreadSheet = new Spreadsheet();
-            $spreadSheet->getActiveSheet()->getDefaultColumnDimension()->setWidth(20);
-            $spreadSheet->getActiveSheet()->fromArray($products);
-            $Excel_writer = new Xls($spreadSheet);
-            header('Content-Type: application/vnd.ms-excel');
-            header('Content-Disposition: attachment;filename="products.xls"');
-            header('Cache-Control: max-age=0');
-            ob_end_clean();
-            $Excel_writer->save('php://output');
-            exit();
-        } catch (Exception $e) {
-            return;
-        }
-    }
-}
