+++ app/Http/Controllers/ProfileController.php
@@ -74,9 +74,6 @@ class ProfileController extends Controller
         ]);
     }
 
-    /**
-     * Delete the user's account.
-     */
     public function destroy(Request $request): RedirectResponse
     {
         $request->validateWithBag('userDeletion', [
@@ -91,7 +88,6 @@ class ProfileController extends Controller
 
         $request->session()->invalidate();
         $request->session()->regenerateToken();
-
 
         return redirect()
             ->to('/');
